requirements.txt
numpy
pillow
streamlit
ultralytics
